Fix segfault when reading from las into multiple buffers
In the presence of multiple buffers (such as the temporary buffers used
by the Mosaic stage) caching the dimensions when reading las doesn't
work: dimensions may be in different locations in different buffers.
The previously used buffers may even be deallocated, in which case the
cached dimensions will be stale pointers, leading to segfaults.

The fix here simply removes dimension caching entirely.  With any
reasonable chunk size this shouldn't lead to performance degradation.