Provide switch to turn on Vagrant's ssh forwarding
The switch is an environment variable, PDAL_VAGRANT_SSH_FOWARD_AGENT,
which, when set to any value, turns on ssh.forward_agent in the
Vagrantfile. More information about ssh configuration options can be
found in the [Vagrant
docs](http://docs.vagrantup.com/v2/vagrantfile/ssh_settings.html).

To turn on ssh forwading automatically when you boot your PDAL Vagrant
VM, put the following in your bash profile:

  export PDAL_VAGRANT_SSH_FORWARD_AGENT=true

It seems to me as if environment variables (with a PDAL_VAGRANT_ prefix)
might be the cleanest way to provide per-developer configuration for the
Vagrantfile. Other suggestions for these customizations are welcome.