Allow faux::Reader to report 0 for getNumPoints
This will be used for simulating readers that have unknown point counts.

This is implemented by a optional boolean argument to one of the faux
reader's constructors. This argument forces the faux reader to report
zero for getNumPoints(), even thought it in fact will provide the number
of points as requested in the second argument.