Build and test bpf without zlib
zlib support was baked into parts of the bpf driver, via the
`BpfSeqIterator::inflate` method and the logic in
`BpfSeqIterator::readBlock`. This patch only defines `inflate` if we
have zlib support, throws a runtime error if the bpf iterator tries to
read a block without zlib, and disables zlib tests if there's no zlib
support.

I'm not sure that my treatment of `readBlock` is correct, as for all I
know there could be a way to read a block without zlib. However, since I
don't know better, I kept things dumb and just threw an error if we
don't have zlib. @abellgithub can we do better?