Build with clang and g++-4.8 on travis
In order to build with clang, we have to compromise a bit and build with
embedded boost. Due to c++11 changes, boost libraries built with g++ can
cause issues when building with clang, if by issues I mean segfaults.
Therefore, we build with embedded boost when using clang.

Clang c++11 support does require updated libstdc (on Travis) so we
install g++-4.8 regardless of our compiler choice.

Since it was easy, this patch also updates our g++ version to 4.8.