Add number_of_returns option to FauxReader
This option tells the FauxReader to write ReturnNumber and
NumberOfReturns to the buffer. NumberOfReturns will be the value passed
in via `number_of_returns` option. ReturnNumber will increase
monotonically from 1 to `number_of_returns`, then reset to 1 and repeat
as necessary.

This patch enables easy testing of high return counts.