Skip, don't fail, when writing high return num las
This patch adds a new option to the las writer,
`discard_high_return_numbers`. This is a boolean option that defaults to
false. If this value is true, the las writer will check the return
number and return counts on incoming points. If the return number is
above the LAS maximum of five returns, then the point is dropped.
Additionally, the numberOfReturns is clipped to five.

This is to support data streams that might have more than five returns
for a given pulse.