Guard a PCLConversions include
This passed Travis b/c stubs are now ON for Travis
(f85550e6d3532460a18847b7f178e552961e3953). I'm not sure if it's
worth the hassle to add a whole new row to the build matrix to test with
stubs off, but a stubs off test would have been necessary to catch this
problem.