Delay filename requirement until processOptions
The las writer previously opened up the output stream on writer
construction, which mandated that the options passed to Writer(Options)
needed to include a "filename" option. This patch defers the "filename"
requirement until `processOptions`.

In order to maintain compatibility with existing code, this patch keeps
the Writer(ostream) and Writer(Options, ostream) signatures. This means
that a las Writer created with an ostream will *not* require a
"filename" argument, and will in fact ignore any filename argument when
constructing the ostream. This is consistent with past behavior.

This patch also changes the protected member variable m_streamManager on
Writer from a OutputStreamManager to a unique_ptr<OutputStreamManager>.
This is to support OutputStreamManager construction *after* Writer
construction.