add PCLVisualizer writer and View kernel
PCLVisualizer writer can be used as an endpoint to spawn a simple
viewer. To use the PCLVisualizer, PCL must be built with
BUILD_visualization=ON and WITH_VTK=ON, and PCL must be enabled within
PDAL.

PDAL_BUNDLE=ON creates the pdal app as application bundle, which is not
explicitly required, but enables some PCL keyboard/mouse interactions in
PCLVisualizer. As reported by Kitware
(http://vtk.1045678.n5.nabble.com/Keyboard-and-mouse-interaction-on-OSX-td5719379.html).

The View kernel allows us to create a simple viewer with only an input
filename, e.g., pdal view input.las.

The writer can also be invoked through existing kernels by passing a
dummy filename with the .pclviz extension. This is a bit of a hack, and
perhaps we should consider adding a "visualize" option that both writes
the requested output format and spawns the viewer.