Port SBET Reader and Writer
This changes PDAL_ADD_DRIVER slightly. Instead of propagating
PDAL_TARGET_OBJECTS to PARENT_SCOPE within the macro, we pass them back
to the driver CMakeLists, where they can be combined with other object
libraries and propagating to PARENT_SCOPE. This is necessary for drivers
that may have both a reader and a writer, in which case the final
library was the only one being passed back up to PARENT_SCOPE with the
original macro.