Specify languages in cmake project
If languages aren't specified, they *should* default to CXX and C, but
on a cmake 3.0.2 build on Red Hat the lack of explicit languages caused
cmake to not find my local gcc/g++ install and instead use the system
compilers. Very weird, I know, but this fixed the problem, I guess
there's no harm in explicitly stating the languages?