Fix NULL pointer in `pdal translate --polygon`
The crop_stage was being initialized to a (zero literal) NULL pointer,
and was never updated to a real Stage *. This led to a segfault when
trying to crop to a polygon with `pdal translate` (and probably would
cause segfaults in other situations as well).

Rather than creating a StageFactory in a third place in the same method,
I just made a method-wide stage factory — losing possible efficiency in
favor of simplicity.

This patch does not include any unit tests, since AFICT there aren't any
unit tests for kernels.