Load only one plugin when guessing plugin names
In `PluginManager::guessLoadByPath`, the driverName was not being
entirely used when loading plugins. Only the plugin type (filter,
reader, writer) was used. This patch checks the actual name of the
driver (e.g. hexbin) against the filename of the library, and only loads
the plugin if the name matches.

Since a plugin only needs to be loaded once, the return value of the
`loadByPath` call is returned from `guessLoadByPath`. If no plugins are
found, the `guessLoadByPath` now returns -1.