Fix failure on PgpointcloudWriterTest
If Postgres is available at runtime but not properly configured for database
creation, or pointcloud extension not available, the tests fail.
Here we sheat by skipping the tests in case of those errors.
It would be good if the infrastructure had a cleaner way of skipping tests
in case a requirement isn't met at runtime.