Check point view size in `getFieldAs`
`PointView::getFieldInternal` uses `deque::operator[]` to map index
values to raw point ids. This means that a user can silently access
beyond the end of of the `m_index` deque with an ill-advised
`getFieldAs` call, leading to either garbage results or a segfault.

Per
[discussion](https://github.com/gadomski/PDAL/commit/c1d54e56e2de841d37f2a1b1c218ed723053f6a9#commitcomment-14415138)
with @abellgithub, this should be considered a user error, albeit one
that we can protect against when in debug mode. Therefore, we use an
assert to check the bounds inside `getFieldAs`.