Add PointXYZ output to writers.pcd
New option writers.pcd.xyz (bool default false) can be used for
writing PointXYZ PCD files. Default behavior remains the same,
so backwards compatibility is retained.

It would be worthwhile, but non-trivial, to make a more robust option
such as output_dims. The problem with using output_dims is the
necessity to make a huge number of point types and the concerns with
memory alignment that could come with that. It might not be too
challenging to implement options for the default PointT provided
with PCL.