Add GDAL_INCLUDE_DIR to interface include dirs
If GDAL is installed to a non-standard location, anything that uses
`PDALUtils.hpp` will fail b/c it can't find `cpl_port.h`. Since we use
`include_directories` instead of `target_include_directories`, the CMake
library interface is not populated with upstream include directories.
This patch fixes that *for GDAL only*.

A more comprehensive population of the target include interface is
probably wise. IMO the best way to be to switch all usages of
`include_directories` to `target_include_directories`, which is some
work.