Improve quality of derivative writer outputs
First, we had a bug in how we computed slope for the D8 method. We were
dividing the slope in the cardinal directions by sqrt(2) instead of the
diagonals.

Second, we offer a switch to output slopes in degrees, not just percent. Note,
this is really just a convenience as analysts should have been able to convert
in the GIS tool of their choice.

Finally, our DSM generation approach had a tendency to leave voids in the max
elevation matrix. We switched to something very similar to what the GDAL writer
does, that is, specifying a radius and considering all points within the
cylinder centered at the current grid cell. At the same time, we remove the DSM
cleaning step.

Future work should consider unifying the behavior of GDALGrid/GDALWriter with
DerivativeWriter such that we can leverage the existing implementation and hole
filling capabilities. We may also be able to convert this to support streaming.