Accelerate morphological operations
filters.pmf and filters.smrf both spend a majority of their time in the
morphological operations (erosion and dilation). As it turns out, we don't gain
much by using Eigen to implement these either. By storing our matrices as
vectors in column-major order, and refactoring the erosion and dilation
functions slightly, we have achieved speedups of greater than 2x for both
filters.

We also had a bug in SMRF, as the erosion and dilation steps were reversed in
the net cutting procedure.

Finally, we have added some additional tests for both erosion and dilation.