Updating outlier, PMF, and SMRF filters
It is desirable to exclude noise points from further processing, as well as
anything but last returns (i.e., return number equals number of returns).
Although this can be accomplished with various other filters, we currently
require that those points be dropped from the PointView if we wanted to exclude
them from the bare earth filters. Provide a mechanism for segmenting last
returns and ignoring an optional, user-defined DimRange. Apply this to both
SMRF and PMF. This allows us to set them aside for the filtering step, without
losing them altogether.

This also marks the first point at which we remove the "extract" and "classify"
options from the PMF and outlier filters. Ground/noise points will always be
classified, but not extracted. Users wishing to extract ground or noise returns
should use the range filter as a downstream processing step.

The outlier filter now allows the user to specify the class they would like to
use for noise points. The default value for noise points has been changed from
18 to 7.

Finally, the outlier filter now uses online updates when computing mean,
variance, and standard deviation.