Provide for normal flipping
User can always set a preferred viewpoint by providing X, Y, and Z coordinates.
If a viewpoint coordinate is not provided, it will be replaced as follows.

* X and Y will default to the midpoint of the X/Y extents
* Z will default to the maximum Z plus 1000

Normals are flipped when the dot product of the normal and the vector between
the current point and the viewpoint is less than zero.