Improve stage overview documentation (#1880)
Convert from bulleted TOC stage listings to definition lists, whereby the
filter name and link are shown as the "term" and a brief stage description is
provided as the definition. The intent is to provide a simpler overview of our
growing list of stages.

Filters are also broken into some categories that attempt to capture what they
do to the data (like creating new dimensions, culling points, etc.).