Configure targets to link libxml2 only if found
Although according to the docs libxml2 is an optional dependency, main
target unconditionally uses LIBXML2_INCLUDE_DIR and LIBXML2_LIBRARIES
which are flagged as -NOTFOUND. That is not valid value for target_*
commands.
This makes libxml2 a required dependency leading to CMake configuration
errors, if not found.

Remove Ilvis2Reader.cpp from sources if libxml2 if not found.