Enable Unicode support for paths on Windows (#2135)
This is a minimal changeset that enables PDAL to recognise filesystem
locations that require Unicode encoded paths (e.g. Japanese paths).

On Windows, every path in ASCII or UTF-8 is converted to UTF-16 and
passed to Win32 API functions or C++ file streams (Microsoft extension)
which expect paths encoded in wchar_t-character string.