Add native PCD I/O (removing reader/writer from PCL plugin)
- Supports reading and writing ASCII
- Supports reading and writing binary
- Readers are streamable, while writers are not
- Does NOT yet support binary compressed
- Removes need for PcdCommon header
- Adds rudimentary tests for PCD reader/writer

A note about the PCD writer. Similar to other writers, the PCD writer can
output a subset of the dimensions if keep_unspecified is set to false. In that
case, only the dimensions specified in the order parameter are written in the
output. The dimension type ("F", "I", or "U") and size (in bytes) can also be
specified, along with precision in the case of ASCII output.

This native driver for PCD will remove a long-standing issue with the previous
PCL-backed drivers, which required a conversion to a fixed, single-precision
point schema, that resulted in loss of extra dimensions (beyond XYZRGBA) and
confusion surrounding subtraction of offsets to avoid loss of precision for
certain coordinate systems.