Limit VoxelCentroidNearestNeighbor results to points within voxel (#2703)
* Limit VoxelCentroidNearestNeighbor results to points within voxel

Previously, VoxelCentroidNearestNeighbor could return points outside the
current voxel. This happened because we search a global KDIndex for the nearest
neighbor of the voxel centroid, which could in fact come from a neighboring
voxel.

The refactored implementation ensures that we are only considering neighbors
within the current voxel when determining which neighbors is nearest the
centroid.

Furthermore, this refactor also addresses some special cases, namely voxels
containing only one point (it simply passes through), and those containing two
points which by definition are equidistant to the centroid. In this case, we
revert to the VoxelCenterNearestNeighbor approach and select the point that is
closest to the voxel center.

* Avoid errors on Windows with max

* Doc update only

Document rules for how points are selected within each voxel.

Remove reference to filters.voxelgrid, which no longer exists.

* Do not calculate bounds, just use first point as offset