filters.colorization keep points outside raster when streaming (#3391)
always return true from ColorizationFilter::processOne regardless of if
the raster read for the x, y position of the point did not succeed. the
output band(s) of points outside the raster are ignored.

the normal mode output and stream mode output should now be equal for
filters.colorization.

added a quick unit test to check this scenario.