Apply initial transform prior to demeaning
It is unlikely that a user will have pre-computed the centroid of the
fixed points and determined an initial guess of the final transformation
with the origin at the centroid. The application of the initial
transform is therefore moved prior to demeaning the moving PointView by
the fixed PointView centroid.