Reverse transform update math
- Little to no impact, but conceptually correct
- The solved transformation, T, is the adjustment necessary after
applying the final_transformation. If we were to use T and
final_transformation to adjust the points, the ordering would be T *
final_transformation * points. Even though T is first as we read
left-to-right, final_transformation is applied to the points first, and
then T (which is correct). We need to match that order when we update
final_transformation: T should be to the left of final_transformation.