Add ICP scale option, update composed transformation with initial guess
- Added a scale argument (default 'false') for a single scale factor to
be included in the ICP transformation solution.
- Added an update to the 'composed_transformation' matrix to include the
'initial_transformation' if one was supplied. This may not be the best
solution, but unless we add yet another transformation matrix to the
metadata, it seemed like the logical place for the actual, final
transformation that was applied to the 'moving' point cloud view.
- Updated the ICP doc with the scale option and edited a few other option
descriptions for consistency/clarity.
- Added a test to IcpFilterTest for recovering scale.