Generalize filters.transformation to all homographies. (#3467)
Include the normalizing step in filters.transformation to allow general
homographies and update the documentation.  All the examples in the
documentation are unaffected (since the last row of the matrix is
[0,0,0,1]).  Homographies allow (painfully!), perspective images of a
point cloud to be created by following the transformation filter by
writers.gdal.  This setup is not ideal, because the effective camera
matrix isn't stored with the the output image.