Use mambaforge and cache conda packages in CI (#3767)
* Have CI use mambaforge instead of miniconda

This implies the use of conda-forge and the use of the mamba dependency
solver.

* Specify default shell of bash -l {0}

Instead of specifying bash -l {0} on each section, we can specify that
as the default shell for the whole CI config

* Try caching conda packages

Setup caching of conda packages by OS, date and cache number.  The cache
number is there to provide the ability to clear the cache by setting
the value to anything other than 0

* Add CI debug task after build and before test

All the debug task does is list all conda installed packages and their
associated versions, this will be helpful in cases where a change in a
dependency causes a breakage.

Co-authored-by: Ognyan Moore <ogi@hobu.co>