writers.ogr: add attr_dims option
Enable populating OGR attributes with one or more dimension values, by
setting the 'attr_dims' option to a comma-separated list, or the special
value 'all'.

If a measure_dim is specified it will be excluded from the attributes.
And X,Y,Z dimensions will always be excluded.

The multicount option (>1) is incompatible with attr_dims and will raise
an error.