writers.ogr: don't set OGR FID (#3865)
* writers.ogr: don't set OGR feature FID

OGR FIDs need to be unique, but if a pipeline includes a merge filter
pointIds may not be. OGR will generate a unique sequence anyway, and
the user can always pass pointId as a dimension to create an attribute.

* writers.ogr: Reset FID correctly between features

Otherwise, GDAL < 3.5 re-uses the same FID for all features.

Also remove one more call to SetFID() when multi-point features are
being used.