Fix writers.las not preserving classFlags to PDRF3
Fixes #3878.

When running a pipeline which reads from a PDRF 6+ source,
and then writing to a PDRF < 6 file, PDAL was not preserving the flag
bits correctly into the classification field.

This change fixes the issue. I also added a warning when writing any
classification values over 31, since those are not supported by
PDRF < 6.

Previously PDAL was preserving such values, so it is *possible*
someone could be relying on that behaviour. I have added a warning if
this happens and also set the classification value to 1 (unclassified)
since that seemed the most sensible value of the choices available.

Since this is only an issue if the user is downgrading the PDRF and
*also* using user-defined or reserved classification values, this
situation should hopefully be very rare.