Initial Workshop Changes (#4054)
* Added jq Conda Install

Added instructions on how to install jq through Conda. jq is used within the workshop, so giving the user the code for the install seemed important.

* Fixed decimal point precision

The example states that "we only have two decimal places of precision on the bounding box", yet showed more than two decimal places. I removed the extra decimals to match user output.

* Update colorization.rst

Added an explanation on how to colorize the point cloud when using plas.io

* Update denoising.rst

Corrected dataformat_id's explanation. Changed the explained value from 3 to the used value of 0.

Updated Fugro Viewer link to their 2022 page.

* Update thinning.rst

Removed an image of repeated, uncopiable code.

* Update dtm.rst

Removed line about invoking a command that is not present or needed

Removed an image that is identical to the one that comes before, only differing in name.

* Update meshing.rst

Fixed broken link to meshing filters.

* Update batch-processing.rst

Included an additional source for the `xargs` program.

* Update capstone.rst

Added `ept.json` to `"filename":"http://na-c.entwine.io/dublin/",` in the provided .json text so that the user can properly obtain the Dublin Entwine data.

* denoising.rst Shortened URL

Shortened the URL to take up less space

* jq install instructions moved down in metadata.rst

* added colorized image back to dtm.rst to be updated