Fix getting/setting the pointFormat of LAZ-files via LasHeader
Until now, 'LasHeader' was accessing the 'pointFormatBits'
of the internal 'las::Header' directly, when getting/setting
the 'pointFormat'. This prevented the 'CompressionMask' from
being applied for compressed LAS-files, resulting in the
'pointFormat' being offset by '0x80'.

The fix simply forwards the getter/setter calls of 'LasHeader'
to the corresponding functions in 'las::Header', which apply
the 'CompressionMask' correctly.